api/requirements.txt
# FastAPI core
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
pydantic-settings==2.9.1

# Database
sqlalchemy==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.15.2

# Auth
python-jose[cryptography]==3.4.0

# Task queue
dramatiq[redis]==1.17.1
redis==5.3.0

# Storage
boto3==1.38.0

# HTTP client
httpx==0.28.1

# OpenAI (shared with bot)
openai==1.76.0

# Presentation generation (shared with bot)
python-pptx==1.0.2
Pillow==11.1.0
requests==2.32.3

# Utils
python-multipart==0.0.20
environs==14.1.1
